Revert "Basic entity testing and some fixes"
This reverts commit c8840c3a8c16cc4813083153acabf46362110fc5.
