--- .travis.yml
@@ -1,12 +1,0 @@
-language: php
-
-php:
-  - 5.5
-  - 5.6
-  - 7
-
-install:
-  - composer self-update
-  - composer install --prefer-dist --no-interaction
-
-script: phpunit
