+++ composer.json
@@ -19,9 +19,6 @@
         "php": ">=5.5.0",
         "guzzlehttp/guzzle": "~6.0"
     },
-    "require-dev": {
-        "phpunit/phpunit": "~4.0"
-    },
     "autoload": {
         "psr-4": {
             "Picqer\\": "src/Picqer"
